Avoid importing chinese_whispers to source version/license (#5)
In order to single-source the version and license information,
setup.py imported the chinese_whispers package.  However,
if installed as a library, the networkx package is not yet available,
and this causes `pip install` to fail.

By directly reading the chinese_whispers/__init__.py file,
one can source the package and license information without first
needing to import the package.

This is the style of single-sourcing recommended by PyPA
(see https://packaging.python.org/guides/single-sourcing-package-version/).